LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting/report-only-unsafe-eval-expected.txt


PASS Eval is allowed because the CSP is report-only
FAIL SPV event is still raised assert_unreached: SPV event has not been received Reached unreachable code
FAIL Violation report status OK. undefined is not an object (evaluating 'data[0]["body"]')
